move console classes to Console directory
