+++ tests/Etc/Console/ConsoleKernel.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace Stancl\Tenancy\Tests\Etc;
+namespace Stancl\Tenancy\Tests\Etc\Console;
 
 use Orchestra\Testbench\Foundation\Console\Kernel;
