+++ tests/Etc/Console/ExampleQuestionCommand.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Stancl\Tenancy\Tests\Etc;
+namespace Stancl\Tenancy\Tests\Etc\Console;
 
 use Illuminate\Console\Command;
